requirements.txt
opencv-python~=4.4.0.46
numpy~=1.19.3
future~=0.18.2
matplotlib~=3.2.2
scikit-image~=0.17.2
Pillow~=8.0.1
torch~=1.7.0+cu110
simpleitk~=2.0.0rc2.dev910+ga138e
tqdm~=4.54.0
sklearn~=0.0
scikit-learn~=0.24.2
torchvision~=0.8.1+cu110
PyYAML~=5.3.1
scipy~=1.5.4
ptflops~=0.6.6
h5py~=3.1.0
tensorboardX~=2.1